+++ frontend/src/components/chart/ChartOptions.jsx
