+++ frontend/src/components/layout/Header.jsx
@@ -5,7 +5,7 @@ import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import MenuIcon from '@material-ui/icons/Menu';
-import TradingViewPrices from 'material/components/misc/TradingViewPrices';
+import TradingViewPrices from 'components/misc/TradingViewPrices';
 import HeaderBG from 'img/headerbg.png';
 import Hidden from '@material-ui/core/Hidden';
