+++ frontend/src/components/misc/SymbolsListSelector.jsx
