+++ frontend/src/components/misc/TradingViewPrices.jsx
