+++ frontend/src/pages/CalculatorPage.jsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import PageContainer from 'material/components/layout/PageContainer';
+import PageContainer from 'components/layout/PageContainer';
 import Typography from '@material-ui/core/Typography';
 import TextField from '@material-ui/core/TextField';
 import { makeStyles } from '@material-ui/core/styles';
